refactor: move DgsDataLoaderOptionsProvider out from internal
